build-requirements.txt
yt-dlp==2024.8.6
pillow==10.4.0
customtkinter==5.2.2
CTkMessagebox==2.7
pyinstaller==6.10.0
ffmpeg-python==0.2.0
buildozer==1.5.0